[fix] 커뮤니티 작성 수정
