+++ src/components/community/community_writing/ImageInputBox/ImageBox/ImageBox.styles.tsx
@@ -1,22 +1,17 @@
 
-import { COLORS } from '../../../../styles/constants/colors';
 import styled from '@emotion/styled';
-interface IconBoxProps {
+
+interface ImageBoxProps {
   width?: number;
   height?: number;
+  alt:string;
   margin?:number;
-  backgroundColor?:string;
 }
 
-export const IconBox = styled.div<IconBoxProps>`
+export const ImageBox = styled.img<ImageBoxProps>`
   width: ${(props) => (props.width ? `${props.width}px` : '100px')};
   height: ${(props) => (props.height ? `${props.height}px` : '100px')};
+  alt: ${(props) => (props.alt ? `이미지 번호 ${props.alt}` : 'noneIndex')};
   margin-right:${(props) => (props.margin ? `${props.margin}px` : '10px')};
   border-radius: 10px;
-  background-color: ${(props) => props.backgroundColor || COLORS.grayscale.Gray4}; // 배경색 스타일 추가
- 
-
 `;
-
-
-// onChange 이벤트 핸들러 추가
