+++ src/components/community/community_writing/ImageInputBox/ImageInputBox.tsx
