+++ src/components/community/community_writing/ImageInputBox/TitleBox/TitleBox.styles.tsx
@@ -1,5 +1,5 @@
 
-import { COLORS } from '../../../../styles/constants/colors';
+import { COLORS } from '../../../../../styles/constants/colors';
 import styled from '@emotion/styled';
 import React from 'react'; // React import 추가
