+++ src/components/community/community_writing/MainForm/CloseIcon/CloseIcon.tsx
