+++ src/components/community/community_writing/MainForm/MainForm.styles.tsx
@@ -1,4 +1,4 @@
-import { COLORS } from '../../../styles/constants/colors';
+
 import styled from '@emotion/styled';
 
 // 가독성을 위해 스타일 파일은 별도로 둡니다.
