+++ src/components/community/community_writing/MainForm/SubSelectBox/SelectListBox./SelectListBox.styles.tsx
@@ -1,4 +1,4 @@
-import { COLORS } from '../../../../../styles/constants/colors';
+
 import styled from '@emotion/styled';
 
 interface SelectListBoxProps {
