+++ src/components/community/community_writing/MainForm/SubSelectBox/SelectListBox./SelectListBox.tsx
@@ -1,4 +1,4 @@
-import { NextPage } from 'next';
+
 import React from 'react';
 
 interface SelectListBoxProps {
