--- src/pages/comm_wrighting.stories.tsx
@@ -1,14 +1,0 @@
-import type { Meta, StoryObj } from '@storybook/react';
-
-import CommunityWriting from './comm_wrighting';
-
-const meta: Meta<typeof CommunityWriting> = {
-  title: 'CommunityWriting',
-  component: CommunityWriting,
-};
-
-export default meta;
-type Story = StoryObj<typeof CommunityWriting>;
-
-export const Primary: Story = {
-};
